feat: driver app UI with BatchOrder, status updates, End of Day + Android project regeneration

+++ apps/driver_app/lib/core/constants.dart
@@ -15,5 +15,5 @@ class AppConstants {
 
   static const int locationUpdateIntervalSeconds = 120;
   static const int syncRetryIntervalSeconds = 30;
-  static const int mapZoomDefault = 14.0;
+  static const double mapZoomDefault = 14.0;
 }
